refactor: hash utils
