+++ src/utils/abi/decodeErrorResult.ts
@@ -17,9 +17,9 @@ import type { Hex } from '../../types/misc.js'
 import type { IsNarrowable, UnionEvaluate } from '../../types/utils.js'
 import { slice } from '../data/slice.js'
 import {
-  type GetFunctionSelectorErrorType,
-  getFunctionSelector,
-} from '../hash/getFunctionSelector.js'
+  type ToFunctionSelectorErrorType,
+  toFunctionSelector,
+} from '../hash/toFunctionSelector.js'
 import {
   type DecodeAbiParametersErrorType,
   decodeAbiParameters,
@@ -59,7 +59,7 @@ export type DecodeErrorResultErrorType =
   | AbiErrorSignatureNotFoundErrorType
   | DecodeAbiParametersErrorType
   | FormatAbiItemErrorType
-  | GetFunctionSelectorErrorType
+  | ToFunctionSelectorErrorType
   | ErrorType
 
 export function decodeErrorResult<const TAbi extends Abi | readonly unknown[]>(
@@ -73,7 +73,7 @@ export function decodeErrorResult<const TAbi extends Abi | readonly unknown[]>(
   const abi_ = [...(abi || []), solidityError, solidityPanic]
   const abiItem = abi_.find(
     (x) =>
-      x.type === 'error' && signature === getFunctionSelector(formatAbiItem(x)),
+      x.type === 'error' && signature === toFunctionSelector(formatAbiItem(x)),
   )
   if (!abiItem)
     throw new AbiErrorSignatureNotFoundError(signature, {
