+++ src/utils/abi/decodeEventLog.test.ts
@@ -1,7 +1,7 @@
 import { assertType, describe, expect, test } from 'vitest'
 
-import { getEventSelector } from '../../utils/hash/getEventSelector.js'
 import { getAddress } from '../address/getAddress.js'
+import { toEventSelector } from '../hash/toEventSelector.js'
 
 import { decodeEventLog } from './decodeEventLog.js'
 
@@ -951,7 +951,7 @@ test('errors: invalid bool', () => {
       data: '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
       eventName: 'Transfer',
       topics: [
-        getEventSelector('Transfer(address,address,bool)'),
+        toEventSelector('Transfer(address,address,bool)'),
         '0x000000000000000000000000d8da6bf26964af9d7eed9e03e53415d37aa96045',
         '0x000000000000000000000000f39fd6e51aad88f6f4ce6ab8827279cfffb92266',
       ],
