+++ src/utils/hash/hashDefinition.test.ts
@@ -1,72 +1,75 @@
 import { expect, test } from 'vitest'
 
-import { hashAbiItem, hashFunction } from './hashFunction.js'
+import { hashDefinition } from './hashDefinition.js'
 
 test('hashes functions', () => {
-  expect(hashFunction('Transfer(address,address,uint256)')).toEqual(
+  expect(hashDefinition('function ownerOf(uint256)')).toEqual(
+    '0x6352211e6566aa027e75ac9dbf2423197fbd9b82b9d981a3ab367d355866aa1c',
+  )
+  expect(hashDefinition('Transfer(address,address,uint256)')).toEqual(
     '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
   )
   expect(
-    hashFunction(
+    hashDefinition(
       'Transfer(address indexed from, address indexed to, uint256 amount)',
     ),
   ).toEqual(
     '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
   )
   expect(
-    hashFunction(
+    hashDefinition(
       'event Transfer(address indexed from, address indexed to, uint256 amount)',
     ),
   ).toEqual(
     '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
   )
-  expect(hashFunction('drawNumber()')).toEqual(
+  expect(hashDefinition('drawNumber()')).toEqual(
     '0xd80ffb20d597d029eb14b9def3d14da7e6d862943d830906185b1b0b576d8f26',
   )
-  expect(hashFunction('drawNumber( )')).toEqual(
+  expect(hashDefinition('drawNumber( )')).toEqual(
     '0xd80ffb20d597d029eb14b9def3d14da7e6d862943d830906185b1b0b576d8f26',
   )
   expect(
-    hashFunction(
+    hashDefinition(
       'ProcessedAccountDividendTracker(uint256,uint256,uint256,uint256,bool,uint256,address)',
     ),
   ).toEqual(
     '0x4a73985b7c9415b88fbbfbb5e2fb377c08586d96f5c42646ecef7e3717587f6a',
   )
   expect(
-    hashFunction(
+    hashDefinition(
       'ProcessedAccountDividendTracker(uint256 indexed foo, uint256 indexed bar, uint256 baz, uint256 a, bool b, uint256 c, address d)',
     ),
   ).toEqual(
     '0x4a73985b7c9415b88fbbfbb5e2fb377c08586d96f5c42646ecef7e3717587f6a',
   )
-  expect(hashFunction('BlackListMultipleAddresses(address[], bool)')).toEqual(
+  expect(hashDefinition('BlackListMultipleAddresses(address[], bool)')).toEqual(
     '0x170cd84eddb1952bf41adcce9be0e44b66ff38f07cddda1cf64d32708742bd2d',
   )
-  expect(hashFunction('checkBatch(bytes)')).toEqual(
+  expect(hashDefinition('checkBatch(bytes)')).toEqual(
     '0x9b6f373667d9cf576e3a17e6aa047c5d864fcb7f41836b11613215db446698d8',
   )
   expect(
-    hashFunction(
+    hashDefinition(
       'Approval(address indexed owner, address indexed approved, uint256 indexed tokenId)',
     ),
   ).toBe('0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925')
   expect(
-    hashFunction(
+    hashDefinition(
       'ApprovalForAll(address indexed owner, address indexed operator, bool approved)',
     ),
   ).toBe('0x17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31')
-  expect(hashFunction('function balanceOf(address owner)')).toBe(
+  expect(hashDefinition('function balanceOf(address owner)')).toBe(
     '0x70a08231b98ef4ca268c9cc3f6b4590e4bfec28280db06bb5d45e689f2a360be',
   )
-  expect(hashFunction('function ownerOf(uint256 tokenId)')).toBe(
+  expect(hashDefinition('function ownerOf(uint256 tokenId)')).toBe(
     '0x6352211e6566aa027e75ac9dbf2423197fbd9b82b9d981a3ab367d355866aa1c',
   )
 })
 
 test('hashes `AbiFunction`', () => {
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'drawNumber',
       type: 'function',
       inputs: [],
@@ -78,7 +81,7 @@ test('hashes `AbiFunction`', () => {
   )
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'BlackListMultipleAddresses',
       type: 'function',
       inputs: [
@@ -93,7 +96,7 @@ test('hashes `AbiFunction`', () => {
   )
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'checkBatch',
       type: 'function',
       inputs: [{ name: 'bytes', type: 'bytes' }],
@@ -105,7 +108,7 @@ test('hashes `AbiFunction`', () => {
   )
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'balanceOf',
       type: 'function',
       inputs: [{ name: 'owner', type: 'address' }],
@@ -115,7 +118,7 @@ test('hashes `AbiFunction`', () => {
   ).toBe('0x70a08231b98ef4ca268c9cc3f6b4590e4bfec28280db06bb5d45e689f2a360be')
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'ownerOf',
       type: 'function',
       inputs: [{ name: 'tokenId', type: 'uint256' }],
@@ -127,7 +130,7 @@ test('hashes `AbiFunction`', () => {
 
 test('hashes `AbiEvent`', () => {
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'Transfer',
       type: 'event',
       inputs: [
@@ -141,7 +144,7 @@ test('hashes `AbiEvent`', () => {
   )
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'ProcessedAccountDividendTracker',
       type: 'event',
       inputs: [
@@ -159,7 +162,7 @@ test('hashes `AbiEvent`', () => {
   )
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'Approval',
       type: 'event',
       inputs: [
@@ -171,7 +174,7 @@ test('hashes `AbiEvent`', () => {
   ).toBe('0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925')
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'ApprovalForAll',
       type: 'event',
       inputs: [
@@ -183,7 +186,7 @@ test('hashes `AbiEvent`', () => {
   ).toBe('0x17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31')
 
   expect(
-    hashAbiItem({
+    hashDefinition({
       name: 'BlackListMultipleAddresses',
       type: 'event',
       inputs: [
