+++ src/utils/hash/toEventSelector.test.ts
@@ -1,60 +1,60 @@
 import { expect, test } from 'vitest'
 
-import { getEventSelector } from './getEventSelector.js'
+import { toEventSelector } from './toEventSelector.js'
 
 test('creates event signature', () => {
-  expect(getEventSelector('Transfer(address,address,uint256)')).toEqual(
+  expect(toEventSelector('Transfer(address,address,uint256)')).toEqual(
     '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
   )
   expect(
-    getEventSelector(
+    toEventSelector(
       'Transfer(address indexed from, address indexed to, uint256 amount)',
     ),
   ).toEqual(
     '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
   )
   expect(
-    getEventSelector(
+    toEventSelector(
       'event Transfer(address indexed from, address indexed to, uint256 amount)',
     ),
   ).toEqual(
     '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
   )
-  expect(getEventSelector('drawNumber()')).toEqual(
+  expect(toEventSelector('drawNumber()')).toEqual(
     '0xd80ffb20d597d029eb14b9def3d14da7e6d862943d830906185b1b0b576d8f26',
   )
-  expect(getEventSelector('drawNumber( )')).toEqual(
+  expect(toEventSelector('drawNumber( )')).toEqual(
     '0xd80ffb20d597d029eb14b9def3d14da7e6d862943d830906185b1b0b576d8f26',
   )
   expect(
-    getEventSelector(
+    toEventSelector(
       'ProcessedAccountDividendTracker(uint256,uint256,uint256,uint256,bool,uint256,address)',
     ),
   ).toEqual(
     '0x4a73985b7c9415b88fbbfbb5e2fb377c08586d96f5c42646ecef7e3717587f6a',
   )
   expect(
-    getEventSelector(
+    toEventSelector(
       'ProcessedAccountDividendTracker(uint256 indexed foo, uint256 indexed bar, uint256 baz, uint256 a, bool b, uint256 c, address d)',
     ),
   ).toEqual(
     '0x4a73985b7c9415b88fbbfbb5e2fb377c08586d96f5c42646ecef7e3717587f6a',
   )
   expect(
-    getEventSelector('BlackListMultipleAddresses(address[], bool)'),
+    toEventSelector('BlackListMultipleAddresses(address[], bool)'),
   ).toEqual(
     '0x170cd84eddb1952bf41adcce9be0e44b66ff38f07cddda1cf64d32708742bd2d',
   )
-  expect(getEventSelector('checkBatch(bytes)')).toEqual(
+  expect(toEventSelector('checkBatch(bytes)')).toEqual(
     '0x9b6f373667d9cf576e3a17e6aa047c5d864fcb7f41836b11613215db446698d8',
   )
   expect(
-    getEventSelector(
+    toEventSelector(
       'Approval(address indexed owner, address indexed approved, uint256 indexed tokenId)',
     ),
   ).toBe('0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925')
   expect(
-    getEventSelector(
+    toEventSelector(
       'ApprovalForAll(address indexed owner, address indexed operator, bool approved)',
     ),
   ).toBe('0x17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31')
@@ -62,7 +62,7 @@ test('creates event signature', () => {
 
 test('creates event signature for `AbiEvent`', () => {
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'Transfer',
       type: 'event',
       inputs: [
@@ -76,7 +76,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'Transfer',
       type: 'event',
       inputs: [
@@ -90,7 +90,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'drawNumber',
       type: 'event',
       inputs: [],
@@ -100,7 +100,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'drawNumber',
       type: 'event',
       inputs: [],
@@ -110,7 +110,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'ProcessedAccountDividendTracker',
       type: 'event',
       inputs: [
@@ -128,7 +128,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'ProcessedAccountDividendTracker',
       type: 'event',
       inputs: [
@@ -146,7 +146,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'BlackListMultipleAddresses',
       type: 'event',
       inputs: [
@@ -159,7 +159,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'checkBatch',
       type: 'event',
       inputs: [{ name: '', type: 'bytes', indexed: false }],
@@ -169,7 +169,7 @@ test('creates event signature for `AbiEvent`', () => {
   )
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'Approval',
       type: 'event',
       inputs: [
@@ -181,7 +181,7 @@ test('creates event signature for `AbiEvent`', () => {
   ).toBe('0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925')
 
   expect(
-    getEventSelector({
+    toEventSelector({
       name: 'ApprovalForAll',
       type: 'event',
       inputs: [
@@ -193,7 +193,7 @@ test('creates event signature for `AbiEvent`', () => {
   ).toBe('0x17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31')
 
   expect(
-    getEventSelector({
+    toEventSelector({
       anonymous: false,
       inputs: [
         {
