+++ src/utils/index.ts
@@ -342,13 +342,21 @@ export {
   defineFormatter,
 } from './formatters/formatter.js'
 export {
-  type GetEventSelectorErrorType,
-  getEventSelector,
-} from './hash/getEventSelector.js'
+  type ToEventSelectorErrorType,
+  toEventSelector,
+  /** @deprecated – use `ToEventSelectorErrorType`. */
+  type ToEventSelectorErrorType as GetEventSelectorErrorType,
+  /** @deprecated – use `toEventSelector`. */
+  toEventSelector as getEventSelector,
+} from './hash/toEventSelector.js'
 export {
-  type GetFunctionSelectorErrorType,
-  getFunctionSelector,
-} from './hash/getFunctionSelector.js'
+  type ToFunctionSelectorErrorType,
+  toFunctionSelector,
+  /** @deprecated – use `ToFunctionSelectorErrorType`. */
+  type ToFunctionSelectorErrorType as GetFunctionSelectorErrorType,
+  /** @deprecated – use `toFunctionSelector`. */
+  toFunctionSelector as getFunctionSelector,
+} from './hash/toFunctionSelector.js'
 export { type IsHashErrorType, isHash } from './hash/isHash.js'
 export { type Keccak256ErrorType, keccak256 } from './hash/keccak256.js'
 export { type Sha256ErrorType, sha256 } from './hash/sha256.js'
